+++ packages/graph/src/core/transaction/useCommitTransaction.ts
@@ -1,7 +1,5 @@
-import type {
-  CoreEdge,
-  CoreNode,
-} from '../../../../graph-core-infra/src/types.ts';
+import { CoreEdge, CoreNode } from '@magic/graph-core-infra/types';
+
 import { createEmptyPayload } from './createEmptyPayload.ts';
 import type {
   CommitTransaction,
