+++ packages/graph/src/core/types.ts
@@ -1,9 +1,7 @@
+import { CoreEdge, CoreNode } from '@magic/graph-core-infra/types';
+
 import type { ComputedRef, Ref } from 'vue';
 
-import type {
-  CoreEdge,
-  CoreNode,
-} from '../../../graph-core-infra/src/types.ts';
 import { GraphSettings } from '../settings/index.ts';
 import { CoreGraphHelpers } from './helpers/types.ts';
 import { NodePositionStoreControls } from './positions/types.ts';
